Thi
Control

   - merged sync_type into phase_type
   - added CommBlock::index_cycle_phase_
   - added initial_exit to control
   - merged CommBlock::q_[phase] with CommBlock::p_[phase]
        (still potentially useful for viewing projections output)
   - added CommBlock::p_output_end() so Output phase ends in CommBlock scope
   - added CommBlock::control_next() to replace CommBlock::control_sync(phase,sync)
   - added cycle_phase[] and cycle_sync[] arrays to define control for control_next()

Adapt

   - fixed bug where doneInserting was not called in first cycle
   - moved adapt_again check from adapt_exit to adapt_end
   - fixed adapt_step_ increment error in call to Factor::create_block

Refresh

   - added warnings to CommBlock::refresh_begin_() to check consistency of is_leaf() with std::vector children_
   - fixed output format error in ERROR

Boundary

   - adding mask support to BoundaryValue UNFINISHED / UNTESTED

Utilities

   - added parse_diff.sh to parse "hg diff" output into an org-mode file

+++ src/Cello/simulation_Simulation.cpp
@@ -365,11 +365,11 @@ void Simulation::initialize_data_descr_() throw()
   // groups
 
   int num_fields = config_->field_group_list.size();
-  for (size_t index_field=0; index_field<num_fields; index_field++) {
+  for (int index_field=0; index_field<num_fields; index_field++) {
     std::string field = config_->field_list[index_field];
     int num_groups = config_->field_group_list[index_field].size();
    
-    for (size_t index_group=0; index_group<num_groups; index_group++) {
+    for (int index_group=0; index_group<num_groups; index_group++) {
       std::string group = config_->field_group_list[index_field][index_group];
       field_descr_->groups()->add(field,group);
     }
